backend/requirements.txt
fastapi
uvicorn
redis
celery
litellm
tenacity
google-cloud-aiplatform
pydantic-settings
supabase
pyjwt[crypto]
